Move check_input_stems under pyani_plus.workflows
This is a precursor to removing the remainder of pyani_plus/snakemake/
