+++ pyani_plus/snakemake/snakemake_scheduler.py
@@ -32,31 +32,6 @@ from pyani_plus import workflows
 from pyani_plus.tools import ToolExecutor
 
 
-def check_input_stems(indir: str) -> dict[str, Path]:
-    """Check input files against approved list of extensions.
-
-    If duplicate stems with approved extensions are present
-    raise a ValueError.
-    """
-    extensions = [".fasta", ".fas", ".fna"]
-    stems = [_.stem for _ in Path(indir).glob("*") if _.suffix in extensions]
-
-    if len(stems) == len(set(stems)):
-        input_files = {
-            _.stem: _ for _ in Path(indir).glob("*") if _.suffix in extensions
-        }
-    else:
-        duplicates = [
-            item for item in stems if stems.count(item) > 1 and item in set(stems)
-        ]
-        msg = (
-            f"Duplicated stems found for {sorted(set(duplicates))}. Please investigate."
-        )
-        raise ValueError(msg)
-
-    return input_files
-
-
 class SnakemakeRunner:
     """Execute Snakemake workflows for pairwise comparisons."""
