+++ pyani_plus/workflows/__init__.py
@@ -20,3 +20,30 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 """Module providing snakemake workflows."""
+
+from pathlib import Path
+
+
+def check_input_stems(indir: str) -> dict[str, Path]:
+    """Check input files against approved list of extensions.
+
+    If duplicate stems with approved extensions are present
+    raise a ValueError.
+    """
+    extensions = [".fasta", ".fas", ".fna"]
+    stems = [_.stem for _ in Path(indir).glob("*") if _.suffix in extensions]
+
+    if len(stems) == len(set(stems)):
+        input_files = {
+            _.stem: _ for _ in Path(indir).glob("*") if _.suffix in extensions
+        }
+    else:
+        duplicates = [
+            item for item in stems if stems.count(item) > 1 and item in set(stems)
+        ]
+        msg = (
+            f"Duplicated stems found for {sorted(set(duplicates))}. Please investigate."
+        )
+        raise ValueError(msg)
+
+    return input_files
