+++ pyani_plus/workflows/snakemake_dnadiff.smk
@@ -19,9 +19,9 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-from pyani_plus.snakemake import snakemake_scheduler
+from pyani_plus.workflows import check_input_stems
 
-indir_files = snakemake_scheduler.check_input_stems(config["indir"])
+indir_files = check_input_stems(config["indir"])
 
 
 def get_genomeA(wildcards):
