+++ tests/snakemake/test_snakemake_anim_workflow.py
@@ -35,6 +35,7 @@ import pytest
 from pyani_plus.private_cli import log_configuration, log_genome, log_run
 from pyani_plus.snakemake import snakemake_scheduler
 from pyani_plus.tools import ToolExecutor, get_delta_filter, get_nucmer
+from pyani_plus.workflows import check_input_stems
 
 from . import compare_matrices
 
@@ -122,9 +123,7 @@ def test_snakemake_rule_filter(  # noqa: PLR0913
     # Record the FASTA files in the genomes table _before_ call snakemake
     log_genome(
         database=db,
-        fasta=list(
-            snakemake_scheduler.check_input_stems(config_anim_args["indir"]).values()
-        ),
+        fasta=list(check_input_stems(config_anim_args["indir"]).values()),
     )
     assert db.is_file()
