+++ tests/snakemake/test_snakemake_fastani_workflow.py
@@ -36,6 +36,7 @@ import pytest
 from pyani_plus.private_cli import log_configuration, log_genome, log_run
 from pyani_plus.snakemake import snakemake_scheduler
 from pyani_plus.tools import ToolExecutor, get_fastani
+from pyani_plus.workflows import check_input_stems
 
 from . import compare_matrices
 
@@ -98,9 +99,7 @@ def test_snakemake_rule_fastani(  # noqa: PLR0913
     # Remove the output directory to force re-running the snakemake rule
     shutil.rmtree(fastani_targets_outdir, ignore_errors=True)
 
-    input_fasta = list(
-        snakemake_scheduler.check_input_stems(config_fastani_args["indir"]).values()
-    )
+    input_fasta = list(check_input_stems(config_fastani_args["indir"]).values())
 
     # Assuming this will match but worker nodes might have a different version
     fastani_tool = get_fastani()
